requirements.txt
boto3==1.34.0
filelock==3.16.0
fsspec==2024.9.0
intel-openmp==2021.4.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
pillow==10.4.0
python-dotenv==1.0.0
sympy==1.13.2
tbb==2021.13.1
torch==2.4.1
torchvision==0.19.1
typing_extensions==4.12.2
